Revert "Fixminors"
