+++ assets/css/dashboard.css
@@ -211,33 +211,7 @@ textarea {
 }
 
 
-.capture{
-    position: absolute;
-    bottom: -12px;
-    right: -8px;
-    padding: 6px;
-    border: 2px solid var(--color-primary-1);
-    width: 45px;
-    height: 45px;
-}
-.img-fluid{
-  object-fit: cover !important;
-}
 
-
-/*loader*/
-
-#loader-div{
-  display: none;
-  text-align: center;
-  background: #fff;
-  border-radius: 9px;
-  position: absolute;
-  padding:20px;
-  top: 50%;
-  left: 50%;
-  z-index: 999;
-}
 
 /* Large devices (desktops, 992px and up) */
 @media (max-width: 992px) {
