+++ functions.php
@@ -88,10 +88,6 @@ function assets($path)
   echo home() . "assets/" . $path;
 }
 
-function themeAssets($template,$path)
-{
-  echo home() . "views/templates/".$template."/assets/" . $path;
-}
 
 function view($fileName)
 {
