+++ views/dashboard/wedding/basic-details.php
@@ -1,37 +1,13 @@
 <?php
-
 locked(['user', 'host', 'manager', 'admin']);
 require('views/partials/dashboard/head.php');
 require('views/partials/dashboard/sidebar.php');
 
-DB::connect();
-$languages = enumToArray(DB::select('information_schema.COLUMNS', 'COLUMN_TYPE', "TABLE_NAME = 'weddings' AND COLUMN_NAME = 'lang'", 'COLUMN_TYPE DESC')->fetch()[0]);
-DB::close();
-
-
-sort($languages);
 
 controller("Wedding");
-controller("Gallery");
-
 $wedding = new Wedding();
 $weddingData = $wedding->getWedding($_REQUEST['id'], $_REQUEST['lang']);
 $story = json_decode($weddingData['story'], true);
-
-
-function getImgURL($name){
-	$gallery = new Gallery();
-	$row=$gallery->getGalleryImg($_REQUEST['id'],$name);
-	
-	if($row['imageURL']){
-		return $row['imageURL'];
-	}
-	else{
-		return false;
-	}
-	
-}
-
 
 ?>
 
@@ -39,132 +15,95 @@ function getImgURL($name){
 
 </head>
 <!--Main Start-->
-<main class="col-md-9 ms-sm-auto col-lg-10 px-md-4 pt-5">
+<main class="col-md-9 ms-sm-auto col-lg-10 px-md-4">
 
-
-		<h1 class="h2">Basic Details</h1>
-
-	<form method="post" id="form" name="createWedding" class="form-wedding" enctype="multipart/form-data">
+	<form method="post" name="updateWedding" class="form-wedding">
 
 		<?php
-	
 		if (isset($_POST['btn-submit'])) {
+			$updateWedding = $wedding->update($_REQUEST['id'], $_REQUEST['lang'], $_REQUEST);
 
-				controller("AWSBucket");
-				$awsObj=new AWSBucket();
-
-				$_REQUEST['host'] = App::getUser()['userID'];
-				$groomArray=array();
-				$brideArray=array();
-				$addToGalleryBride=array();
-				$addToGalleryGroom=array();
-
-				$groomArray['weddingID']=$_REQUEST['id'];
-				$brideArray['weddingID']=$_REQUEST['id'];
-
-				// upload bride img to aws bucket
-				if(isset($_FILES['bride']['name']) && strlen($_FILES['bride']['name'])>0){
-					$gallery = new Gallery();
-
-					$uploadedURL = $awsObj->uploadToAWS($_FILES,'bride');
-					if($uploadedURL['error']){
-						echo '<div class="alert alert-danger">'.$uploadedURL['errorMsg'].'</div>';
-					}
-					else{					
-						$brideArray['imageURL'] = $uploadedURL['url'];	
-						$brideArray['imageName']='bride';
-						$brideArray['type']='bride';
-					}
-					$addToGalleryBride = $gallery->update($brideArray);
-					
-				}
-
-				 // upload groom img to aws bucket
-				if(isset($_FILES['groom']['name']) && strlen($_FILES['groom']['name'])>0 ){
-					$gallery = new Gallery();
-
-					$uploadedURL = $awsObj->uploadToAWS($_FILES,'groom');
-					if($uploadedURL['error']){
-						echo '<div class="alert alert-danger">'.$uploadedURL['errorMsg'].'</div>';
-					}
-					else{					
-						$groomArray['imageURL'] = $uploadedURL['url'];	
-						$groomArray['imageName']='groom';
-						$groomArray['type']='groom';
-					}
-					$addToGalleryGroom = $gallery->update($groomArray);
-				}
-
-				$_REQUEST['weddingID']=$_REQUEST['id'];
-		
-				$createWedding = $wedding->update($_REQUEST['id'],$_REQUEST['lang'],$_REQUEST);
-				
-				if ($createWedding['error']) {
-					?>
-					<div class="alert alert-danger">
-						<?php
-						foreach ($createWedding['errorMsgs'] as $msg) {
-							if (count($msg))
-								echo $msg[0] . "<br>";
-						}
-						?>
-					</div>
+			if ($updateWedding['error']) {
+				?>
+				<div class="alert alert-danger">
 					<?php
-				}else if ($addToGalleryBride['error']) {
+					foreach ($updateWedding['errorMsgs'] as $msg) {
+						if (count($msg))
+							echo $msg[0] . "<br>";
+					}
 					?>
-					<div class="alert alert-danger">
-						<?php
-						foreach ($addToGalleryBride['errorMsgs'] as $msg) {
-							if (count($msg))
-								echo $msg[0] . "<br>";
-						}
-						?>
-					</div>
-					<?php
-				}else if ($addToGalleryGroom['error']) {
-					?>
-					<div class="alert alert-danger">
-						<?php
-						foreach ($addToGalleryGroom['errorMsgs'] as $msg) {
-							if (count($msg))
-								echo $msg[0] . "<br>";
-						}
-						?>
-					</div>
-					<?php
-				} else{
-					redirect("wedding/" . $_REQUEST['weddingID'] . "/" . $_REQUEST['lang'] . "/our-story");
-				}
+				</div>
+				<?php
+			} else
+				redirect("wedding/" . $_REQUEST['id'] . "/" . $_REQUEST['lang'] . "/hosts");
 
-			}
+		}
 
 		?>
+		<h1 class="h2">Our Story</h1>
+		<div class="row">
 
-			<div class="row text-center">
+			<input type="text" hidden name="basic-details">
+			<!-- How we met -->
+			<div class="mb-3 col-8">
+				<label for="howWeMet" class="form-label">How We Met</label>
+				<textarea class="form-control" id="howWeMet" name="howWeMet"
+					placeholder="Enter how Bride & Groom met for the first time"
+					rows="3"><?= $_REQUEST['howWeMet'] ?? str_replace("<br>", "\r\n", $story['howWeMet']) ?></textarea>
+			</div>
 
+			<!-- When we met -->
+			<div class="mb-3 col-4">
+				<label for="whenWeMet" class="form-label">When We Met</label>
+				<select id="whenWeMet" name="whenWeMet" class="form-control">
+					<option hidden>Select Year</option>
+					<?php
+					for ($i = 1990; $i <= date('Y'); $i++):
+						?>
+						<option value="<?= $i ?>" <?php if ($story['whenWeMet'] == $i)
+							echo "selected" ?>>
+							<?= $i ?>
+						</option>
+						<?php
+					endfor;
+					?>
+				</select>
+			</div>
 
-				<!-- groom pic -->
-			    <div class="col-sm-6">
-			      <label for="groom" class="form-label" style="position: relative;">
+			<!-- Engagement -->
+			<div class="mb-3 col-8">
+				<label for="engagement" class="form-label">Engagement</label>
+				<textarea class="form-control" id="engagement" name="engagement"
+					placeholder="Enter how Bride & Groom got engaged"
+					rows="3"><?= $_REQUEST['engagement'] ?? str_replace("<br>", "\r\n", $story['engagement']) ?></textarea>
+			</div>
 
-			      	Groom Photo<br>
-			      	    <img id="groomImage" src="<?php if(getImgURL('groom')){echo getImgURL('groom');}else{ echo assets('img/upload.png');} ?>" alt="Groom Image" class="rounded-circle" style="width: 150px; height: 150px;">
+			<!-- Engagement Year -->
+			<div class="mb-3 col-4">
+				<label for="engagementYear" class="form-label">Engagement Year</label>
+				<select id="engagementYear" name="engagementYear" class="form-control">
+					<option hidden>Select Year</option>
+					<?php
+					for ($i = 1990; $i <= date('Y'); $i++):
+						?>
+						<option value="<?= $i ?>" <?php if ($story['engagementYear'] == $i)
+							echo "selected" ?>>
+							<?= $i ?>
+						</option>
+						<?php
+					endfor;
+					?>
+				</select>
+			</div>
 
-			      	    <span class="btn btn-sm btn-secondary capture"><i class="fas fa-camera"></i></span>
+			<!-- Memorable Moments -->
+			<div class="mb-3 col-12">
+				<label for="memorableMoments" class="form-label">Memorable Moments</label>
+				<textarea class="form-control" id="memorableMoments" name="memorableMoments"
+					placeholder="Add any sweet memorable moments you like to share"
+					rows="3"><?= $_REQUEST['memorableMoments'] ?? str_replace("<br>", "\r\n", $story['memorableMoments']) ?></textarea>
+			</div>
 
-			      </label>
-
-			    	<input type="file" class="form-control" id="groom" accept="image/*" name="groom"  onchange="displayGroomImage(this)">
-			    </div>
-
-     	<!--  bride pic -->
-			    <div class="col-sm-6">
-			      <label for="bride" class="form-label" style="position: relative;">
-
-			      	Bride Photo<br>
-			      	    <img id="brideImage" src="<?php if(getImgURL('bride')){echo getImgURL('bride');}else{ echo assets('img/upload.png');} ?>" alt="Bride Image" class="rounded-circle" style="width: 150px; height: 150px;">
-
-			      	    <span class="btn btn-sm btn-secondary capture"><i class="fas fa-camera"></i></span>
 
 			<h1 class="h2 mt-3">Basic Details</h1>
 			<!-- Groom Qualifications -->
@@ -198,69 +137,7 @@ function getImgURL($name){
 			</div>
 
 
-			      </label>
-
-			      <input type="file" class="form-control" id="bride" accept="image/*" name="bride"  onchange="displayBrideImage(this)" >
-
-			    </div>
-  		
-
-				<!-- Groom Name -->
-				<div class="mb-3 col-sm-6">
-					<label for="groomName" class="form-label">Groom Name</label>
-					<input type="text" class="form-control" id="groomName" name="groomName"
-						placeholder="Enter Groom Name" value="<?= $_REQUEST['groomName'] ?? $weddingData['groomName'] ?>">
-				</div>
-
-				<!-- Bride Name -->
-				<div class="mb-3 col-sm-6">
-					<label for="brideName" class="form-label">Bride Name</label>
-					<input type="text" class="form-control" id="brideName" name="brideName"
-						placeholder="Enter Bride Name" value="<?= $_REQUEST['brideName'] ?? $weddingData['brideName'] ?>">
-				</div>
-
-				<!-- From (Bride/Groom) -->
-				<div class="mb-3 col-sm-6">
-					<label class="form-label" for="fromRole">From</label>
-
-					<select class="form-select" id="fromRole" name="fromRole">
-						<option value="bride" <?= ($weddingData['fromRole'] == 'bride') ? 'selected' : '' ?>>Bride</option>
-						<option value="groom" <?= ($weddingData['fromRole'] == 'groom') ? 'selected' : '' ?>>Groom</option>
-
-					</select>
-				</div>
-
-				<!-- Language -->
-				<div class="mb-3 col-sm-6">
-					<label for="lang" class="form-label">Language</label>
-					<select class="form-select" id="lang" name="lang">
-						<?php foreach ($languages as $lang) {
-							?>
-							<option value="<?= $lang ?>" <?php
-							  if ($weddingData['lang'] == $lang)
-								  echo 'selected';
-							  elseif ($lang == 'en')
-								  echo 'selected' ?>>
-								<?= Locale::getDisplayLanguage($lang, "en") ?>
-							</option>
-							<?php
-						} ?>
-					</select>
-				</div>
-				<!-- Wedding Name -->
-				<div class="mb-3 col-sm-6">
-					<label for="weddingName" class="form-label">Wedding Name</label>
-					<input type="text" class="form-control" id="weddingName" name="weddingName"
-						placeholder="Thota vaari pelli sandhadi" value="<?= $_REQUEST['weddingName'] ?? $weddingData['weddingName'] ?>">
-				</div>
-
-				<!-- Wedding ID -->
-				<div class="mb-3 col-sm-6">
-					<label for="weddingID" class="form-label">Wedding ID</label>
-					<input type="text" class="form-control" id="weddingID" name="weddingID"
-						placeholder="KishoreWedsSwathi" value="<?= $_REQUEST['id'] ?>" readonly disabled>
-				</div>
-			</div>
+		</div>
 
 
 
@@ -270,47 +147,43 @@ function getImgURL($name){
 
 	</div>
 
+	<script>
+		let weddings = []
+		<?php
+		foreach ($weddings as $wedding) {
+			echo "weddings.push('" . $wedding['weddingID'] . "')\n";
+		}
+		?>
+
+		function generateWeddingID(groomName, brideName) {
+			let weddingID = groomName + "Weds" + brideName;
+
+			if (weddings.includes(weddingID)) {
+				weddingID = groomName + "-Weds-" + brideName;
+			}
+
+			return weddingID;
+		}
+
+		function updateWeddingID() {
+			const groomName = document.querySelector('#groomName').value.trim();
+			const brideName = document.querySelector('#brideName').value.trim();
+			const weddingIDField = document.querySelector('#weddingID');
+
+			if (groomName.length && brideName.length) {
+				const newWeddingID = generateWeddingID(groomName, brideName);
+				weddingIDField.value = newWeddingID;
+			} else weddingIDField.value = "";
+		}
+
+		document.querySelector('#groomName').addEventListener('focusout', updateWeddingID);
+		document.querySelector('#brideName').addEventListener('focusout', updateWeddingID);
+		document.querySelector('#groomName').addEventListener('keyup', updateWeddingID);
+		document.querySelector('#brideName').addEventListener('keyup', updateWeddingID);
+	</script>
 </main>
 
-<script type="text/javascript">
-	
-	//display bride img
-			function displayBrideImage(input) {
-			  const file = input.files[0];
 
-			  if (file) {
-			    const reader = new FileReader();
-			    
-			    reader.onload = function (e) {
-			      document.getElementById('brideImage').src = e.target.result;
-
-			    };
-
-			    reader.readAsDataURL(file);
-			  }
-			  
-			}
-
-			// display groom img
-			function displayGroomImage(input) {
-
-			  const file = input.files[0];
-
-			  if (file) {
-			    const reader = new FileReader();
-
-			    reader.onload = function (e) {
-			      document.getElementById('groomImage').src =e.target.result;
-
-			    };
-
-			    reader.readAsDataURL(file);
-			  }
-
-			}
-
-
-</script>
 
 <!--Main End-->
