--- views/templates/Royal-Bramhin/assets/css/app.css
@@ -1,329 +1,0 @@
-
-nav *{
-  user-select: none;
-  color: var(--color-secondary-1) !important;
-}
-/*nav{
-  position: absolute !important;
-  top: 0 !important;
-  left: 0 !important;
-
-}*/
-nav a{
-  font-family: var(--base-font) !important;
-}
-nav .nav-link{
-  font-size: 18px;
-  margin: auto 10px;
-}
-nav .nav-link:hover{
-  opacity: .8;
-}
-
-nav .nav-link:focus,
-nav .nav-link:active{
-  opacity: .6;
-}
-
-nav .btn-secondary{
-  color: var(--color-primary-1) !important;
-  width: 180px !important;
-}
-
-.navbar-toggler{
-  border: none !important;
-  padding: 0 !important;
-}
-.navbar-toggler:focus,
-.navbar-toggler:active{
-  box-shadow: none !important;
-}
-
-.navbar {
-    transition: top 0.5s ease-in-out;
-    position: fixed;
-    width: 100%;
-    top: -120px;
-}
-
-.navbar.visible {
-    top: 0;
-}
-
-#app{
-  /*margin-top: 20vh !important;*/
-}
-
-.navbar-brand{
-  font-size: 24px;
-}
-
-.dropdown-menu *{
-  color: var(--color-black) !important;
-
-}
-
-/* --------------- */
-
-a{
-    text-decoration: none !important;
-}
-
-.fixed-button {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    z-index: 9999;
-}
-
-/* section head*/
-.section-head{
-  color: var(--color-primary-1);
-  text-align: center;
-  margin:1.3rem 0;
-}
-
-section{
-  padding-top: 2.5rem;
-}
-
-/* hero section css */
-
-#heroSection{
-  margin: 0;
-  padding: 0;
-  height: 100vh;
-}
-
-.carousel-caption{
-    top: 50% !important;
-    left: 50% !important;
-    transform: translate(-50%, -50%) !important;
-    right: 0;
-    bottom: 0;
-}
-
-    #heroCarousel {
-      position: relative;
-    }
-
-    #heroCarousel .item {
-      background-size: cover;
-      background-position: center;
-      height: 100vh; /* Set your preferred height */
-      position: relative;
-    }
-
-        #heroCarousel .item h2
-        {
-          font-size: 3rem;
-        }
-
-    .days,.min,.hours,.sec{
-      font-size: 2rem;
-    }
-
-    #heroCarousel .item::before {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background: rgba(0, 0, 0, 0.5); /* Adjust the overlay color and opacity */
-    }
-
-    #heroCarousel .carousel-caption {
-      /*position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-*/      color: white; /* Adjust the text color */
-      text-align: center;
-    }
-
-    .lines {
-    text-align: center;
-    position: relative;
-    color: var(--color-secondary-1);
-    padding: 0 10px;
-}
-
-.lines::before,
-.lines::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    width: 80px; /* Adjust the width as needed */
-    height: 1px;
-    background-color: black; /* Adjust the color as needed */
-}
-
-.lines::before {
-    left: 0;
-    transform: translate(-100%, -50%);
-}
-
-.lines::after {
-    right: 0;
-    transform: translate(100%, -50%);
-}
-
-/* hero section css ends*/
-
-
-/* about section css starts */
-
-.image-column{
-  position: relative;
-}
-.image-column .img-bg{
-     width: 250px;
-    position: absolute;
-    top: -12px;
-    z-index: -1;
-    right: -1px;
-    background-color: var(--color-primary-1);
-    height: 105%;
-    border-bottom-right-radius: 9px;
-    border-top-right-radius: 9px;
-}
-
-    .vertical-lines {
-    text-align: center;
-    position: relative;
-    color: var(--color-secondary-1);
-    padding: 0 10px;
-}
-
-.vertical-lines::before,
-.vertical-lines::after {
-    content: "";
-    position: absolute;
-    top: 50%;
-    width: 3px; /* Adjust the width as needed */
-    height: 20px;
-    background-color: var(--color-secondary-3); /* Adjust the color as needed */
-}
-
-.vertical-lines::before {
-    left: 0;
-    transform: translate(-100%, -50%);
-}
-
-.vertical-lines::after {
-    right: 0;
-    transform: translate(100%, -50%);
-}
-
-/* about section css ends */
-
-/* bride groom css starts */
-
-.brideImg,.groomImg{
-  border-radius: 50%;
-  width: 150px;
-  height: 150px;
-  object-fit: cover;
-}
-
-/* bride groom css ends */
-
-
-/* our story css starts */
- .timeline {
-            position: relative;
-            padding: 20px 0;
-        }
-
-        .timeline::before {
-            content: '';
-            position: absolute;
-            width: 2px;
-            background-color: var(--color-secondary-2); /* Line color */
-            top: 0;
-            bottom: 0;
-            left: 50%;
-            transform: translateX(-50%);
-        }
-
-        .timeline-item {
-            position: relative;
-            margin: 20px 0;
-        }
-
-        .timeline-item::before {
-            content: '';
-            position: absolute;
-            width: 20px;
-            height: 20px;
-            background-color: var(--color-secondary-1); /* Dot color */
-            border-radius: 50%;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-        }
-
-        .timeline-content {
-            width: 50%;
-            background: var(--color-primary-5);
-            padding: 10px 20px;
-            font-size: 14px;
-            position: relative;
-        }
-        .timeline-layout{
-          position: absolute;
-          top: -12px;
-          left: -12px;
-          width: 100%;
-          height: 100%;
-          background-color: var(--color-secondary-5);
-        }
-
-         /* Align odd timeline items to the left */
-        .timeline-item:nth-child(odd) .timeline-content {
-            text-align: right;
-            margin-left: -5%;
-        }
-
-        /* Align even timeline items to the right */
-        .timeline-item:nth-child(even) .timeline-content {
-            text-align: left;
-            margin-left: 55%;
-        }
-
-/* out story css ends */
-
-/* events css start */
-
-  .pink-border {
-    border: 20px solid pink;
-  }
-
-/* event css ends */
-
-/* Large devices (desktops, 992px and up) */
-@media (max-width: 992px) {
-
-}
-
-/* Medium devices (tablets, 768px and up) */
-@media (max-width: 768px) {
-}
-
-
-/* Small devices (Smartphones, 425px and up) */
-@media (max-width: 425px) {
-  .navbar-brand {
-    font-size: 20px;
-  }
-}
-
-/* Extra Small devices (Smartphones, 350px and up) */
-@media (max-width: 350px) {
-  .navbar-brand {
-    font-size: 16px !important;
-  }
-
-}
-
