--- views/templates/Royal-Bramhin/footer.php
@@ -1,4 +1,0 @@
-
-<h4 class="footer mt-3 text-center py-5 fs-6">
-  &copy; 2024 By Rehan & Kavya
-</h4>
